requirements.txt
numpy==2.3.1
pandas==2.3.1
plotly==6.2.0
python-dotenv==1.1.1
requests==2.32.4
streamlit==1.47.1
yfinance==0.2.65
groq